added function reportProject() added calculating shortest possible time to end project added functions to detect which date is earlier or later in contributors.cpp reformatted code to be more readable

+++ structures.h
@@ -16,7 +16,7 @@ struct structureContributors {
 };
 
 struct date {
-    int day, month, year;
+    short day, month, year;
 };
 
 struct taskStructure {
